Create Counting Mini Game
- Create scene for Counting game.

Counter.cs - holds logic to generate number to count to and controls level of difficulty.

FlowerGeneration.cs - Script for counting game that generates the same number of flowers to match the number the player is supposed to count to.

MathEquation.cs - add comments, use built-in assignment to assure the level passed in is valid.

PlayAnimation.cs & AdditionProblem.cs - add comments