Create Timed Challenge and Create Menu in Main Area
Challenges.cs - holds data for a timed challenge so it can be a part of the saved data.

Game.cs - added challenge objects for addition and subtraction.

GlobalControl.cs - add challenge data to the save and load player functions.

MainAreaMenu.cs - controls what each button does in the main menu.

TimedChallenge.cs - controls the logic of the timed challenge and displays the correct screens to the user.